rename contracts constants file
